+++ src/main/java/core/SynthLogic/Effects/NoiseEffect.java
@@ -10,7 +10,7 @@ public class NoiseEffect implements EffectRack{
     public NoiseEffect(){
     }
     @Override
-    public float[] applyEffect(float[] mixBuffer) {
+    public double[] applyEffect(double[] mixBuffer) {
         for(int i = 0; i < mixBuffer.length; i++){
             mixBuffer[i] += Math.min(mixBuffer[i],Math.random()/80)/5;
         }
