+++ src/main/java/core/Visuals/WaveformUpdateListener.java
@@ -1,5 +1,5 @@
 package core.Visuals;
 
 public interface WaveformUpdateListener {
-    void updateWaveform(float[] newWaveform);
+    void updateWaveform(double[] newWaveform);
 }
